Changed PolygonTool. Added rotation and translation to Objects.
Changed the PolygonTool to draw regular polygons with a given number of sides. Added a function called move(x,y) and a function called rotate(angle) to the Object class. move() moves an object by the specified x and y values. Rotate rotates the function by the specified angle (I believe it defaults to radians, but if you change the degree mode, it will adjust).