Added TextTool
Implemented TextTool. Replaced the Title with a Text object and adjusted Extras Panel text accordingly. Renamed fonts to be a little more consistent. Added global inputDelay variable to better manage the input delay in different tools. Added code to suppress normal key input when the TextTool is active and to notify the TextTool when shift is down (checks in the driver because that was easier than trying to detect two keys at once in one keyPressed() function).