Performance Improvements
Basically this stops the side panel from being drawn over along with being drawn to the screen every time Draw() is called. I also added the ability to create a message that will display in the center of the screen until X seconds have passed.

Possible Bugs: If the user somehow draws on the side of the screen that is reserved for the toolbar, then that drawing will draw over the toolbar.