Fixed Text Tool overlap. Minor changes.
Added the welcome message text box back in. Changed the text box to add a line break when it gets to the edge of the screen. FIXED THE TEXT BOX WIDTH. Made the Pencil Tool draw a special scribble object (it's a curve but now I can mess around with it later if I want).